all microservices first working version (+kafka)

+++ rooms-service/internal/models/models.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Room struct {
-	CreatorUID int32     `json:"creatorUID"`
-	RoomID     int32     `json:"roomID"`
+	CreatorUID int64     `json:"creatorUID"`
+	RoomID     int64     `json:"roomID"`
 	Name       string    `json:"name"`
 	IsPrivate  bool      `json:"isPrivate"`
 	CreatedAt  time.Time `json:"createdAt"`
